refactor(strider): search
